--- src/main/java/com/dnd/gooding/oauth/command/domain/CreateOAuthService.java
@@ -1,24 +1,0 @@
-package com.dnd.gooding.oauth.command.domain;
-
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-public class CreateOAuthService {
-
-	private OAuthRepository oAuthRepository;
-
-	public CreateOAuthService(OAuthRepository oAuthRepository) {
-		this.oAuthRepository = oAuthRepository;
-	}
-
-	@Transactional
-	public void createOAuth(OAuth oAuth) {
-		oAuthRepository.findByoAuthId(oAuth.getoAuthId()).ifPresentOrElse(
-			x -> {
-			}, () -> {
-				oAuthRepository.save(oAuth);
-			}
-		);
-	}
-}
