+++ src/main/java/com/dnd/gooding/user/command/application/CreateMemberService.java
@@ -3,8 +3,6 @@ package com.dnd.gooding.user.command.application;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.dnd.gooding.common.model.EmailSet;
-import com.dnd.gooding.common.model.InterestSet;
 import com.dnd.gooding.common.model.UserRole;
 import com.dnd.gooding.oauth.command.domain.OAuthId;
 import com.dnd.gooding.user.command.domain.Member;
@@ -12,11 +10,11 @@ import com.dnd.gooding.user.command.domain.MemberId;
 import com.dnd.gooding.user.command.domain.MemberRepository;
 
 @Service
-public class MemberService {
+public class CreateMemberService {
 
 	public MemberRepository memberRepository;
 
-	public MemberService(MemberRepository memberRepository) {
+	public CreateMemberService(MemberRepository memberRepository) {
 		this.memberRepository = memberRepository;
 	}
